requirements.txt
matplotlib==2.2.2
tensorflow-gpu==1.8.0
pandas==0.23.1
scikit-learn==0.19.1
scipy==1.1.0
xgboost==0.72
tabulate==0.8.2

